Rename smoothing_plugins -> single_waypt_smoothing_plugins
