+++ moveit_core/single_waypt_smoothing_plugins/CMakeLists.txt
@@ -1,3 +1,6 @@
+# Filtering plugins for realtime control.
+# They smooth a single waypoint at a time.
+
 # Plugin implementations
 
 add_library(moveit_butterworth_filter SHARED
