+++ moveit_core/single_waypt_smoothing_plugins/include/moveit/single_waypt_smoothing_plugins/ruckig_filter.h
@@ -33,64 +33,49 @@
  *********************************************************************/
 
 /* Author: Andy Zelenak
-   Description: A first-order Butterworth low-pass filter. There is only one parameter to tune.
+   Description: Jerk-limited smoothing with the Ruckig library.
  */
 
 #pragma once
 
-#include <cstddef>
-
 #include <moveit/robot_model/robot_model.h>
-#include <moveit/smoothing_plugins/smoothing_base_class.h>
+#include <moveit/single_waypt_smoothing_plugins/smoothing_base_class.h>
 
-namespace smoothing_plugins
+#include <ruckig/ruckig.hpp>
+
+namespace single_waypt_smoothing_plugins
 {
 /**
- * Class ButterworthFilter - Implementation of a signal filter to soften jerks.
- * This is a first-order Butterworth low-pass filter. First-order was chosen for 2 reasons:
- * - It doesn't overshoot
- * - Computational efficiency
+ * Class RuckigFilter - Implementation of a signal filter to soften jerks.
  */
-class ButterworthFilter
+class RuckigFilter
 {
 public:
-  /**
-   * Constructor.
-   * @param low_pass_filter_coeff Larger filter_coeff-> more smoothing of servo commands, but more lag.
-   * Rough plot, with cutoff frequency on the y-axis:
-   * https://www.wolframalpha.com/input/?i=plot+arccot(c)
-   */
-  ButterworthFilter(double low_pass_filter_coeff);
+  RuckigFilter();
 
-  double filter(double new_measurement);
+  double filter();
 
-  void reset(const double data);
-
-private:
-  static constexpr std::size_t FILTER_LENGTH = 2;
-  std::array<double, FILTER_LENGTH> previous_measurements_;
-  double previous_filtered_measurement_;
-  // Scale and feedback term are calculated from supplied filter coefficient
-  double scale_term_;
-  double feedback_term_;
+  void reset();
 };
 
 // Plugin
-class ButterworthFilterPlugin : public SmoothingBaseClass
+class RuckigFilterPlugin : public SmoothingBaseClass
 {
 public:
-  ButterworthFilterPlugin(){};
+  RuckigFilterPlugin(){};
 
-  bool initialize(rclcpp::Node::SharedPtr node, const moveit::core::JointModelGroup& group /*unused*/, size_t num_dof,
-                  double /*unused*/) override;
+  bool initialize(rclcpp::Node::SharedPtr node, const moveit::core::JointModelGroup& group, size_t num_dof,
+                  double timestep) override;
 
-  bool doSmoothing(std::vector<double>& position_vector, std::vector<double>& /*unused*/) override;
+  bool doSmoothing(std::vector<double>& /*unused*/, std::vector<double>& velocity_vector) override;
 
   bool reset(const std::vector<double>& joint_positions) override;
 
 private:
-  rclcpp::Node::SharedPtr node_;
-  std::vector<ButterworthFilter> position_filters_;
   size_t num_dof_;
+  rclcpp::Node::SharedPtr node_;
+  std::shared_ptr<ruckig::Ruckig<0, true /*debug*/>> ruckig_;
+  std::shared_ptr<ruckig::InputParameter<0>> ruckig_input_;
+  std::shared_ptr<ruckig::OutputParameter<0>> ruckig_output_;
 };
-}  // namespace smoothing_plugins
+}  // namespace single_waypt_smoothing_plugins
