+++ moveit_core/single_waypt_smoothing_plugins/src/smoothing_base_class.cpp
@@ -1,0 +1,1 @@
+#include <moveit/single_waypt_smoothing_plugins/smoothing_base_class.h>
