--- moveit_core/smoothing_plugins/src/smoothing_base_class.cpp
@@ -1,1 +1,0 @@
-#include <moveit/smoothing_plugins/smoothing_base_class.h>
