+++ moveit_ros/moveit_servo/config/panda_simulated_config.yaml
@@ -26,7 +26,7 @@ publish_joint_velocities: false
 publish_joint_accelerations: false
 
 ## Plugins for smoothing outgoing commands
-smoothing_filter_plugin_name: "smoothing_plugins::RuckigFilterPlugin"
+smoothing_filter_plugin_name: "single_waypt_smoothing_plugins::RuckigFilterPlugin"
 
 ## MoveIt properties
 move_group_name:  panda_arm  # Often 'manipulator' or 'arm'
