+++ moveit_ros/moveit_servo/config/panda_simulated_config_pose_tracking.yaml
@@ -25,7 +25,7 @@ publish_joint_velocities: true
 publish_joint_accelerations: false
 
 ## Plugins for smoothing outgoing commands
-smoothing_filter_plugin_name: "smoothing_plugins::ButterworthFilterPlugin"
+smoothing_filter_plugin_name: "single_waypt_smoothing_plugins::ButterworthFilterPlugin"
 
 ## MoveIt properties
 move_group_name:  panda_arm  # Often 'manipulator' or 'arm'
